Split SaleCampaign component
+ extract SaleLogoViewer
+ extract SaleDescriptionViewer
+ extract SaleStateWidget
+ use refactored select field inside invest form
+ minor fixes
